Bgnd (#61)
* fixing script_functions.c bgnd related functions

* updating function signatures to avoid conflics

* working on bgnd

* arena background element helpers reconstruct

* arena background script elements

* light function signatures from bgnd corrections

* arenas script functions and scenery lights

* background arena weapons

* bettle lair funcitons

* removing duplicated strut typedef

+++ include/runtime/image.h
@@ -174,7 +174,7 @@ void insert_ani_texture_control_item(AniTextureControl* atc, AniTextureControlIt
 AniTextureControl* ck_ani_texture_control_item(AniTextureControlItem* item);
 
 AniTextureControl* append_texture_by_name_to_atomic_material_id(
-    int slot, char* name, RpAtomic* atomic, unsigned short material_id, int flag);
+    int slot, char* name, RpAtomic* atomic, int material_id, int flag);
 AniTextureControl* attach_named_wiff_to_first_material(int slot, char* name, ImageMkSobj* mkobj);
 AniTextureControl* attach_wiff_to_atomic_material(int slot, char* name, RpAtomic* atomic, char* tex_name);
 AniTextureControl* append_wiff_to_clump_material(int slot, char* name, RpClump* clump, char* tex_name);
